show existing config files in list format and allow uplaod from new tab
